+++ org-slipbox-rpc.el
@@ -70,6 +70,7 @@
 (defconst org-slipbox-rpc-method-agenda "slipbox/agenda")
 (defconst org-slipbox-rpc-method-search-refs "slipbox/searchRefs")
 (defconst org-slipbox-rpc-method-capture-node "slipbox/captureNode")
+(defconst org-slipbox-rpc-method-capture-template "slipbox/captureTemplate")
 (defconst org-slipbox-rpc-method-ensure-file-node "slipbox/ensureFileNode")
 (defconst org-slipbox-rpc-method-append-heading "slipbox/appendHeading")
 (defconst org-slipbox-rpc-method-append-heading-to-node "slipbox/appendHeadingToNode")
@@ -189,6 +190,10 @@ When NOCASE is non-nil, use case-insensitive matching."
 (defun org-slipbox-rpc-capture-node (params)
   "Capture a new node using PARAMS."
   (org-slipbox-rpc-request org-slipbox-rpc-method-capture-node params))
+
+(defun org-slipbox-rpc-capture-template (params)
+  "Capture using generic template PARAMS."
+  (org-slipbox-rpc-request org-slipbox-rpc-method-capture-template params))
 
 (defun org-slipbox-rpc-ensure-file-node (file-path title)
   "Ensure FILE-PATH exists as a file node with TITLE."